Feature/valid check

+++ src/main/java/com/example/post/controller/v1/PostController.java
@@ -1,9 +1,13 @@
 package com.example.post.controller.v1;
 
+import com.example.board.service.BoardService;
+import com.example.comment.service.CommentService;
 import com.example.common.dto.PageList;
 import com.example.common.dto.PageRequest;
 import com.example.common.dto.Result;
 import com.example.common.exception.DataNotFoundException;
+import com.example.common.exception.InvalidParameterException;
+import com.example.common.exception.NotAllowedOperationException;
 import com.example.post.dto.PostBody;
 import com.example.post.model.Post;
 import com.example.post.service.PostService;
@@ -12,13 +16,17 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
+
 @RestController
 @RequestMapping(value="/v1/boards/{boardSeq}/posts")
 @RequiredArgsConstructor
 public class PostController {
     final Logger logger = LoggerFactory.getLogger(PostController.class);
 
+    private final BoardService boardService;
     private final PostService postService;
+    private final CommentService commentService;
 
     /**
      * 포스트 목록 조회
@@ -32,7 +40,7 @@ public class PostController {
                                       PageRequest pageRequest) {
         logger.info("getPostList ::: {} {}", boardSeq, pageRequest);
 
-        postService.validateBoardSeq(boardSeq);
+        checkExistence(boardSeq);
 
         return postService.getPostList(pageRequest, boardSeq);
     }
@@ -51,6 +59,8 @@ public class PostController {
                         @PathVariable long postSeq) {
         logger.info("getPost ::: {} {} {}",
                 userSeq, boardSeq, postSeq);
+
+        checkExistence(boardSeq);
 
         Post post = getValidatedPost(userSeq, boardSeq, postSeq);
 
@@ -72,12 +82,16 @@ public class PostController {
         logger.info("deletePost ::: {} {} {}",
                 userSeq, boardSeq, postSeq);
 
+        checkExistence(boardSeq);
+
         Post post = getValidatedPost(userSeq, boardSeq, postSeq);
+
+        if (commentService.isCommentExist(postSeq)) {
+            throw new NotAllowedOperationException();
+        }
 
         return postService.deletePost(post);
     }
-
-
 
 
     /**
@@ -90,11 +104,11 @@ public class PostController {
     @PostMapping(value="")
     public Post createPost(@RequestAttribute long userSeq,
                            @PathVariable long boardSeq,
-                           @RequestBody PostBody body) {
+                           @RequestBody @Valid PostBody body) {
         logger.info("createPost ::: {} {} {}",
                 userSeq, boardSeq, body);
 
-        postService.validateBoardSeq(boardSeq);
+        checkExistence(boardSeq);
 
         Post post = Post.builder()
                 .memberNo(userSeq)
@@ -119,9 +133,11 @@ public class PostController {
     public Post modifyPost(@RequestAttribute long userSeq,
                            @PathVariable long boardSeq,
                            @PathVariable long postSeq,
-                           @RequestBody PostBody body) {
+                           @RequestBody @Valid PostBody body) {
         logger.info("modifyPost ::: {} {} {} {}",
                 userSeq, boardSeq, postSeq, body);
+
+        checkExistence(boardSeq);
 
         Post post = getValidatedPost(userSeq, boardSeq, postSeq);
 
@@ -131,21 +147,31 @@ public class PostController {
         return postService.modifyPost(post);
     }
 
+
+    private void checkExistence(long boardSeq) {
+        boardService.validateBoardSeq(boardSeq);
+    }
+
+
     private Post getValidatedPost(long userSeq, long boardSeq, long postSeq) {
         Post post = postService.getPost(postSeq);
 
-        validatePost(post, userSeq, boardSeq);
+        if (post == null) {
+            throw new DataNotFoundException();
+        }
+
+        checkEquality(post, userSeq, boardSeq);
         return post;
     }
 
 
-    private void validatePost(Post post, long userSeq, long boardSeq) {
+    private void checkEquality(Post post, long userSeq, long boardSeq) {
         if (post.getBoardNo() != boardSeq) {
-            throw new DataNotFoundException();
+            throw new InvalidParameterException();
         }
 
         if (post.getMemberNo() != userSeq) {
-            throw new DataNotFoundException();
+            throw new InvalidParameterException();
         }
     }
 }
